refactor(compositionapi): grid layout fake-arrow
